#76 fixed conflict error when last player wanted to leave the lobby

+++ src/main/java/ch/uzh/ifi/hase/soprafs21/service/LobbyService.java
@@ -174,8 +174,10 @@ public class LobbyService {
 
             //deleting the connector that isn't used anymore
             LobbyConnector connectorToDelete = lobbyConnectorRepository.findByLastLobbyId(lobbyId);
-            lobbyConnectorRepository.delete(connectorToDelete);
-            lobbyConnectorRepository.flush();
+            if(connectorToDelete!=null){
+                lobbyConnectorRepository.delete(connectorToDelete);
+                lobbyConnectorRepository.flush();
+            }
         }
     }
